+++ CHANGES.md
@@ -4,6 +4,7 @@ CHANGES
 ## Version 0.2dev
 
  * Fix trailing newlines becoming spaces in title
+ * Add handler options `markup_prefix` and `markup_suffix`
 
 ## Version 0.1.2 (2013-01-18T05:47:16Z)